requirements.txt
streamlit
langchain
langchain_groq
langchain_community
sentence-transformers
faiss-cpu
python-dotenv
bs4